+++ docs/release-notes.md
@@ -8,6 +8,7 @@ Released on May 5, 2020.
 * Fixed spacing issue in the `Menu` component (thanks @JeroenDeDauw)
 * Added edit icon to the Visual Editor edit action in the navbar (thanks @JeroenDeDauw)
 * Added `hideDiscussionLink` option to the `PageTools` component (thanks @JeroenDeDauw)
+* Added `placeholder` option to the `SearchBar` component (thanks @JeroenDeDauw)
 
 ### Chameleon 2.2.0